requirements.txt
scrapy
pymongo
python-dotenv
pandas
dash
plotly

